Finish register
